feat: user manage permissions

+++ backend/src/users/users.service.ts
@@ -8,30 +8,40 @@ import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
+import { AbilityFactory, Action } from 'src/ability/ability.factory';
 
 @Injectable()
 export class UsersService {
-  constructor(@InjectRepository(User) private userRepo: Repository<User>) {}
+  constructor(
+    @InjectRepository(User) private userRepo: Repository<User>,
+    private abilityFactory: AbilityFactory,
+  ) {}
 
-  async create(createUserDto: CreateUserDto) {
+  async create(createUserDto: CreateUserDto, currentUser: User) {
+    this.abilityFactory.checkPermission(currentUser, Action.Create, User);
+
     const exists = await this.userRepo.findOneBy({ name: createUserDto.name });
     if (exists) throw new BadRequestException('User already exists');
 
     return this.userRepo.save(createUserDto);
   }
 
-  async remove(id: number) {
+  async remove(id: number, currentUser: User) {
+    this.abilityFactory.checkPermission(currentUser, Action.Delete, User);
+
     const user = await this.findById(id);
     return this.userRepo.remove(user);
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
+  async update(id: number, updateUserDto: UpdateUserDto, currentUser: User) {
+    this.abilityFactory.checkPermission(currentUser, Action.Update, User);
     const user = await this.findById(id);
     user.role = updateUserDto.role;
     return this.userRepo.save(user);
   }
 
-  findAll() {
+  findAll(currentUser: User) {
+    this.abilityFactory.checkPermission(currentUser, Action.Read, User);
     return this.userRepo.find();
   }
